Create class_max_elevation_city.py
The City class has the following attributes: name, country (where the city is located), elevation (measured in meters), and population (approximate, according to recent statistics). 

Of the 3 defined instances, the max_elevation_city function compares and returns the name of the city and country (separated by a comma) that meets specified elevation amount and has the highest elevation.